refactors: rename db module to database, drop the only foreign key, filter API-sourced tables
