--- service/db.py
@@ -1,4 +1,0 @@
-from flask.ext.sqlalchemy import SQLAlchemy
-from octopus.core import app
-
-db = SQLAlchemy(app)
