+++ service/initialise.py
@@ -11,7 +11,7 @@ def initialise():
     # Something like:
     # SQLALCHEMY_DATABASE_URI="postgresql+psycopg2://user:pass@localhost/database_name"
 
-    from service.db import db
+    from service.database import db
     from sqlalchemy_utils import database_exists, create_database
 
     if not database_exists(db.engine.url):
@@ -22,5 +22,3 @@ def initialise():
             app.logger.error('Exception {0} while trying to create database specified in config. Make sure the DB user has sufficient rights to create databases, this ability is not usually granted by default. Or create the database manually and rerun the app.'.format(e.message))
     else:
         app.logger.info('Database specified in config already exists, not creating or modifying it during init.')
-
-    from service import models  # initialise SQLAlchemy mappings with our models
