+++ service/lib/util.py
@@ -1,13 +1,16 @@
 import sqlalchemy_utils
-from service.db import db
+from service.database import db, FA_API_TABLES
 
 
 def get_tableobj_by_name(tablename):
-    return db.metadata.tables[tablename]
+    return FA_API_TABLES[tablename]
 
 
 def get_model_class_by_tablename(table):
-    return sqlalchemy_utils.get_class_by_table(db.Model, get_tableobj_by_name(table))
+    mclass = sqlalchemy_utils.get_class_by_table(db.Model, get_tableobj_by_name(table))
+    if not mclass:
+        raise ValueError("Can't find model class for table name %s" % table)
+    return mclass
 
 
 def load(filename):
